Moved pandas backend into separate module
