+++ utils/utils.py
@@ -9,6 +9,7 @@ from collections import OrderedDict
 import psutil
 from tempfile import mkstemp
 
+from .pandas_backend import pb, import_pandas_into_module_namespace
 from utils_base_env.benchmarks import benchmark_mapper
 
 
@@ -39,63 +40,6 @@ ny_taxi_data_files_sizes_MB = OrderedDict(
         "trips_xat.csv": 8600,
     }
 )
-
-
-def init_modin_on_hdk(pd):
-    # Calcite initialization
-    data = {"a": [1, 2, 3]}
-    df = pd.DataFrame(data)
-    df = df + 1
-    _ = df.index
-
-
-def import_pandas_into_module_namespace(namespace, mode, ray_tmpdir=None, ray_memory=None):
-    if mode == "Pandas":
-        print("Pandas backend: pure Pandas")
-        import pandas as pd
-    else:
-        if mode == "Modin_on_ray":
-            import ray
-
-            if not ray_tmpdir:
-                ray_tmpdir = "/tmp"
-            if not ray_memory:
-                ray_memory = 200 * 1024 * 1024 * 1024
-            if not ray.is_initialized():
-                ray.init(
-                    include_dashboard=False,
-                    _plasma_directory=ray_tmpdir,
-                    _memory=ray_memory,
-                    object_store_memory=ray_memory,
-                )
-            os.environ["MODIN_ENGINE"] = "ray"
-            print(
-                f"Pandas backend: Modin on Ray with tmp directory {ray_tmpdir} and memory {ray_memory}"
-            )
-        elif mode == "Modin_on_dask":
-            os.environ["MODIN_ENGINE"] = "dask"
-            print("Pandas backend: Modin on Dask")
-        elif mode == "Modin_on_python":
-            os.environ["MODIN_ENGINE"] = "python"
-            print("Pandas backend: Modin on pure Python")
-        elif mode == "Modin_on_hdk":
-            os.environ["MODIN_ENGINE"] = "native"
-            os.environ["MODIN_STORAGE_FORMAT"] = "hdk"
-            os.environ["MODIN_EXPERIMENTAL"] = "True"
-            print("Pandas backend: Modin on HDK")
-        else:
-            raise ValueError(f"Unknown pandas mode {mode}")
-        import modin.pandas as pd
-
-        # Some components of Modin on HDK engine are initialized only
-        # at the moment of query execution, so for proper benchmarks performance
-        # measurement we need to initialize these parts before any measurements
-        if mode == "Modin_on_hdk":
-            init_modin_on_hdk(pd)
-    if not isinstance(namespace, (list, tuple)):
-        namespace = [namespace]
-    for space in namespace:
-        space["pd"] = pd
 
 
 def get_percentage(error_message):
@@ -751,9 +695,14 @@ def run_benchmarks(
         "query_name",
     ]
 
-    benchmarks = benchmark_mapper
-
-    run_benchmark = __import__(benchmarks[bench_name]).run_benchmark
+    # Set current backend, !!!needs to be run before benchmark import!!!
+    pb.set_backend(
+        pandas_mode=pandas_mode,
+        ray_tmpdir=ray_tmpdir,
+        ray_memory=ray_memory,
+    )
+    
+    run_benchmark = __import__(benchmark_mapper[bench_name]).run_benchmark
 
     parameters = {
         "data_file": data_file,
@@ -768,6 +717,7 @@ def run_benchmarks(
         "validation": validation,
         "extended_functionality": extended_functionality,
     }
+
 
     db_params = DBParams(
         server=db_server, port=db_port, user=db_user, password=db_pass, name=db_name
